+++ tests/mocking.rs
@@ -165,6 +165,94 @@ fn external_mock() {
     assert_eq!(events, ActionEvents::empty());
 }
 
+#[test]
+fn entity_command() {
+    let mut app = App::new();
+    app.add_plugins((MinimalPlugins, EnhancedInputPlugin))
+        .add_input_context::<TestContext>()
+        .finish();
+
+    let context = app
+        .world_mut()
+        .spawn((TestContext, actions!(TestContext[Action::<Test>::new()])))
+        .id();
+
+    let mut actions = app
+        .world_mut()
+        .query::<(&Action<Test>, &ActionState, &ActionEvents)>();
+
+    app.update();
+
+    let (&action, &state, events) = actions.single(app.world()).unwrap();
+    assert!(!*action);
+    assert_eq!(state, ActionState::None);
+    assert!(events.is_empty());
+
+    app.world_mut()
+        .commands()
+        .entity(context)
+        .mock_once::<TestContext, Test>(ActionState::Fired, true);
+
+    // Update once to apply the command, and once to process the actual mock
+    app.update();
+    app.update();
+
+    let (&action, &state, &events) = actions.single(app.world()).unwrap();
+    assert!(*action);
+    assert_eq!(state, ActionState::Fired);
+    assert_eq!(events, ActionEvents::FIRE | ActionEvents::START);
+
+    app.update();
+
+    let (&action, &state, &events) = actions.single(app.world()).unwrap();
+    assert!(!*action);
+    assert_eq!(state, ActionState::None);
+    assert_eq!(events, ActionEvents::COMPLETE);
+}
+
+#[test]
+fn world_entity() {
+    let mut app = App::new();
+    app.add_plugins((MinimalPlugins, EnhancedInputPlugin))
+        .add_input_context::<TestContext>()
+        .finish();
+
+    let context = app
+        .world_mut()
+        .spawn((TestContext, actions!(TestContext[Action::<Test>::new()])))
+        .id();
+
+    let mut actions = app
+        .world_mut()
+        .query::<(&Action<Test>, &ActionState, &ActionEvents)>();
+
+    app.update();
+
+    let (&action, &state, events) = actions.single(app.world()).unwrap();
+    assert!(!*action);
+    assert_eq!(state, ActionState::None);
+    assert!(events.is_empty());
+
+    app.world_mut()
+        .entity_mut(context)
+        .mock_once::<TestContext, Test>(ActionState::Fired, true)
+        .unwrap();
+
+    app.update();
+
+    let (&action, &state, &events) = actions.single(app.world()).unwrap();
+    assert!(*action);
+    assert_eq!(state, ActionState::Fired);
+    assert_eq!(events, ActionEvents::FIRE | ActionEvents::START);
+
+    app.update();
+
+    let (&action, &state, &events) = actions.single(app.world()).unwrap();
+    assert!(!*action);
+    assert_eq!(state, ActionState::None);
+    assert_eq!(events, ActionEvents::COMPLETE);
+}
+
 #[derive(Component)]
 struct TestContext;